refactor(stm): moved the off circuit checks functions to the rolling state
