+++ mithril-stm/src/proof_system/ivc_halo2_snark/prover_input.rs
@@ -16,10 +16,10 @@ use crate::{
     proof_system::ivc_halo2_snark::{
         prover_input_helpers::{
             IvcTransitionType, assert_message_hash_matches_preimage, build_next_accumulator,
-            build_next_state, create_snark_message_for_next_state, verify_certificate_proof,
+            build_next_state, verify_certificate_proof,
         },
         prover_setup::IvcSnarkProverSetup,
-        rolling_state::IvcRollingState,
+        rolling_state::{IvcProverInputChecks, IvcRollingState},
     },
 };
 
@@ -37,51 +37,41 @@ pub(crate) struct IvcProverInput {
     pub(crate) transition_type: IvcTransitionType,
 }
 
-/// Bundles [`IvcProverInput::off_circuit_checks`]'s outputs, carried into
-/// [`IvcProverInput::finish_preparation`] once the certificate proof is available.
-pub(crate) struct IvcProverInputChecks<D: MembershipDigest> {
-    transition_type: IvcTransitionType,
-    certificate_message_hash: MessageHash,
-    certificate_merkle_tree_commitment: MerkleTreeCommitment,
-    message: Vec<u8>,
-    aggregate_verification_key_for_snark: AggregateVerificationKeyForSnark<D>,
-}
-
 impl IvcProverInput {
     /// Runs every `prepare` off circuit check that does not need the certificate proof:
     /// classifies the step, validates the epoch advance and protocol-parameter lookahead
     /// against the rolling state, and checks the message hashes to the supplied preimage.
-    pub(crate) fn off_circuit_checks<D: MembershipDigest>(
-        message: &[u8],
-        aggregate_verification_key_for_snark: &AggregateVerificationKeyForSnark<D>,
-        protocol_message_preimage: &ProtocolMessagePreimage,
-        rolling_state: &IvcRollingState,
-    ) -> StmResult<IvcProverInputChecks<D>> {
-        let transition_type = IvcTransitionType::try_compute_transition_type(
-            rolling_state,
-            protocol_message_preimage,
-        )?;
+    // pub(crate) fn off_circuit_checks<D: MembershipDigest>(
+    //     message: &[u8],
+    //     aggregate_verification_key_for_snark: &AggregateVerificationKeyForSnark<D>,
+    //     protocol_message_preimage: &ProtocolMessagePreimage,
+    //     rolling_state: &IvcRollingState,
+    // ) -> StmResult<IvcProverInputChecks<D>> {
+    //     let transition_type = IvcTransitionType::try_compute_transition_type(
+    //         rolling_state,
+    //         protocol_message_preimage,
+    //     )?;
 
-        rolling_state.assert_correct_parameters(
-            protocol_message_preimage,
-            aggregate_verification_key_for_snark,
-            message,
-            transition_type,
-        )?;
+    //     rolling_state.assert_correct_parameters(
+    //         protocol_message_preimage,
+    //         aggregate_verification_key_for_snark,
+    //         message,
+    //         transition_type,
+    //     )?;
 
-        let (certificate_message_hash, certificate_merkle_tree_commitment) =
-            create_snark_message_for_next_state(aggregate_verification_key_for_snark, message)?;
+    //     let (certificate_message_hash, certificate_merkle_tree_commitment) =
+    //         create_snark_message_for_next_state(aggregate_verification_key_for_snark, message)?;
 
-        assert_message_hash_matches_preimage(certificate_message_hash, protocol_message_preimage)?;
+    //     assert_message_hash_matches_preimage(certificate_message_hash, protocol_message_preimage)?;
 
-        Ok(IvcProverInputChecks {
-            transition_type,
-            certificate_message_hash,
-            certificate_merkle_tree_commitment,
-            message: message.to_vec(),
-            aggregate_verification_key_for_snark: aggregate_verification_key_for_snark.clone(),
-        })
-    }
+    //     Ok(IvcProverInputChecks {
+    //         transition_type,
+    //         certificate_message_hash,
+    //         certificate_merkle_tree_commitment,
+    //         message: message.to_vec(),
+    //         aggregate_verification_key_for_snark: aggregate_verification_key_for_snark.clone(),
+    //     })
+    // }
 
     /// Completes `prepare` once the certificate proof is available: verifies it, advances the
     /// chain state, and folds the accumulator. `checks` must come from [`Self::off_circuit_checks`]
@@ -150,11 +140,10 @@ impl IvcProverInput {
         rolling_state: &IvcRollingState,
         prover_setup: &IvcSnarkProverSetup,
     ) -> StmResult<Self> {
-        let checks = Self::off_circuit_checks(
+        let checks = rolling_state.off_circuit_checks(
             message,
             aggregate_verification_key_for_snark,
             protocol_message_preimage,
-            rolling_state,
         )?;
 
         Self::finish_preparation(
